% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betterKmeans}
\alias{betterKmeans}
\title{Better-K-Means clustering}
\usage{
betterKmeans(inputMatrix, k, epsilon, maxIter, omega, expression, drawPoints)
}
\arguments{
\item{inputMatrix}{N-dimensional matrix with data to clustering}

\item{k}{The number of clusters}

\item{epsilon}{Max squared error to stop clustering}

\item{maxIter}{Max iteration number to stop clustering}

\item{omega}{The number which is weight from 0.0 to 1.0}

\item{expression}{String with math expression,
which includes parameters symbols (B1,B2...BN) and arguments symbols (X1,X2...XN).}

\item{drawPoints}{Matrix with points from specific range,
if you don't plan to plot result you can ignore this parameter}
}
\value{
The instance of betterKmeans class which is list with:

* centers - Center of the each cluster

* labels - Label of each point, which indicates cluster of this point

* errors - Mean error of each cluster (sum of distance divided by number of points in the cluster)

* eigenVectors - Eigen vector of each cluster, calculated from covariance of points in the cluster.

* eigenValues - Eigen value of each cluster, calculated from  covariance of points in the cluster.

* pointsToDraw - Result of function of each cluster with drawPoints as arguments.
If you don't plan to plot result you can ignore this.

* inputMatrix - copy of matrix from function parameters
}
\description{
It's modification traditional K-Means clustering method, which use omega input, cluster bases
and no-linear function interpolation to calculate distances.

It has own plot method.
}
\examples{
library(KM)

## large_letter_G is dataset with 2d representation of the large letter "G"
inputMatrix <- large_letter_G

## "(X1^2)*B1+X1*B2+B3" - Quadratic function
result <- betterKmeans(inputMatrix, 5, 0.01, 10, c(0.5), "(X1^2)*B1+X1*B2+B3", as.matrix(seq(-2, 2, length.out = 400)))
plot(result)
}

