% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeans}
\alias{kmeans}
\title{K-Means clustering}
\usage{
kmeans(inputMatrix, k, epsilon, maxIter)
}
\arguments{
\item{inputMatrix}{N-dimensional matrix with data to clustering}

\item{k}{The number of clusters}

\item{epsilon}{Max squared error to stop clustering}

\item{maxIter}{Max iteration number to stop clustering}
}
\value{
The instance of kmeans class which is list with:

* centers - Center of the each cluster

* labels - Label of each point, which indicates cluster of this point

* errors - Mean error of each cluster (sum of distance divided by number of points in the cluster)

* inputMatrix - copy of matrix from function parameters
}
\description{
It's traditional clustering method without any modifications.

It has own plot method.
}
\examples{
library(KM)

## chinese_father is dataset with 2d representation of the word "Father" in Chinese writing
inputMatrix <- chinese_father

result <- kmeans(inputMatrix, 3, 0.01, 30)
plot(result)
}

