% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kOmeans}
\alias{kOmeans}
\title{K-Omega-Means clustering}
\usage{
kOmeans(inputMatrix, k, epsilon, maxIter, omega)
}
\arguments{
\item{inputMatrix}{N-dimensional matrix with data to clustering}

\item{k}{The number of clusters}

\item{epsilon}{Max squared error to stop clustering}

\item{maxIter}{Max iteration number to stop clustering}

\item{omega}{N-dimensional vector with weights from 0.0 to 1.0}
}
\value{
The instance of kOmeans class which is list with:

* centers - Center of the each cluster

* labels - Label of each point, which indicates cluster of this point

* errors - Mean error of each cluster (sum of distance divided by number of points in the cluster)

* eigenVectors - Eigen vector of each cluster, calculated from covariance of points in the cluster.

* eigenValues - Eigen value of each cluster, calculated from  covariance of points in the cluster.

* inputMatrix - copy of matrix from function parameters
}
\description{
It's modification traditional K-Means clustering method, which use n-dimensional omega input and cluster bases
to calculate distances.

It has own plot method.
}
\examples{
library(KM)

## small_letter_f is dataset with 2d representation of the small letter "f"
inputMatrix <- small_letter_f

result <- kOmeans(inputMatrix, 4, 0.01, 50, c(5.0, 0.5))
plot(result)
}

