% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.betterKmeans}
\alias{plot.betterKmeans}
\title{Plot result of betterKmeans function}
\usage{
\method{plot}{betterKmeans}(result, cex = 0.1, pch = 19, xlab = "X",
  ylab = "Y", asp = 1, xlim = c(0, 1), ylim = c(0, 1), type = NULL,
  main = NULL, sub = NULL)
}
\arguments{
\item{result}{Instance of betterKmeans class}
}
\description{
All elements are in cluster-specific colors.

It plots:

* Cluster points

* Centers of clusters

* Bases of clusters which are triangular areas

* Functions of clusters
}

