% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.kmeans}
\alias{plot.kmeans}
\title{Plot result of kmeans function}
\usage{
\method{plot}{kmeans}(result, cex = 0.1, pch = 19, xlab = "X",
  ylab = "Y", asp = 1, xlim = c(0, 1), ylim = c(0, 1), type = NULL,
  main = NULL, sub = NULL)
}
\arguments{
\item{result}{Instance of kmeans class}
}
\description{
All elements are in cluster-specific colors.

It plots:

* Cluster points

* Centers of clusters
}

