% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.kOmeans}
\alias{plot.kOmeans}
\title{Plot result of kOmeans function}
\usage{
\method{plot}{kOmeans}(result, cex = 0.1, pch = 19, xlab = "X",
  ylab = "Y", asp = 1, xlim = c(0, 1), ylim = c(0, 1), type = NULL,
  main = NULL, sub = NULL)
}
\arguments{
\item{result}{Instance of kOmeans class}
}
\description{
All elements are in cluster-specific colors, except function f(x)=0, which is black.

It plots:

* Cluster points

* Centers of clusters

* Bases of clusters which are triangular areas

* Function f(x)=0
}

